+++ src/main/java/lotto/command/CommandFactory.java
@@ -1,12 +1,6 @@
-package stringcalculator.command;
+package lotto.command;
 
 import java.util.Set;
-
-import stringcalculator.command.AddCommand;
-import stringcalculator.command.Command;
-import stringcalculator.command.DivideCommand;
-import stringcalculator.command.MinusCommand;
-import stringcalculator.command.MultiplyCommand;
 
 public class CommandFactory {
     private static final Set<String> SUPPORTED_OPERATORS = Set.of("+", "-", "*", "/");
