+++ src/test/java/lotto/domain/NumberTest.java
@@ -1,4 +1,4 @@
-package stringcalculator.domain;
+package lotto.domain;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
@@ -7,8 +7,6 @@ import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.CsvSource;
-
-import stringcalculator.domain.Number;
 
 public class NumberTest {
