+++ src/test/java/lotto/domain/StringArithmeticInputTestTest.java
@@ -1,4 +1,4 @@
-package stringcalculator.domain;
+package lotto.domain;
 
 import org.assertj.core.api.Assertions;
 import org.junit.jupiter.api.DisplayName;
@@ -6,9 +6,6 @@ import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.CsvSource;
 import org.junit.jupiter.params.provider.NullAndEmptySource;
-
-import stringcalculator.domain.Number;
-import stringcalculator.domain.StringArithmeticInput;
 
 public class StringArithmeticInputTestTest {
